backend/requirements.txt
django
djangorestframework
django-cors-headers
requests
gunicorn
psycopg2-binary
whitenoise
dj-database-url
python-dotenv
django-cors-headers
